Wire-format bounds checks + dimension overflow guards (audit Phase 2)
Round-2 adversarial audit (D) and round-1 IPC audit (C) flagged
several places where attacker-controlled or hostile-input lengths
could trigger OOB reads, int32 overflow into a too-small buffer,
or signed-negation UB. This change closes those.

P0-D1: xi::ipc::Reader::str() / bytes() — embedded `n = u32()` was
copied / assigned with no bounds check against `end_ - p_`. Outer
recv_frame caps total payload at 16 MiB, but a malformed frame whose
inner str-length declares 4 GB would silently 4 GB-allocate (or
read OOB on the assign path). Add explicit guard mirroring `take()`.

P0-D2: xi_shm.hpp::alloc_image — `int64_t pixels = w*h*ch` was
truncated to int32 in payload_size. A 50000x50000x4 image computes
pixels=1e10, truncates to a much smaller positive int32; the bump
allocator carves out the small block but the caller writes the full
w*h*ch bytes, scribbling past the block end into the next block's
payload. Reject any allocation whose pixel count exceeds INT32_MAX
at the API boundary, plus reject non-positive dims.

D-P1-7: xi_image_pool.hpp::create —
  (a) `(size_t)w * h * ch` cast applies only to the first
      multiplicand; `h * ch` is int32 mul first, signed-overflow UB
      for h*ch > INT32_MAX. Compute pixels in int64 first.
  (b) Live counters (live_count_, total_created_) were incremented
      BEFORE acquire_slot_. On pool exhaustion the counters drifted
      permanently (D-P1-8). Move counter bumps to after slot acquire.
  (c) Cap per-image at 1 GiB to bound pressure from hostile / careless
      plugins. Comfortably above any real CV input.
  (d) `entry->generation = gen` after slots_.store was a use-after-
      release of the unique_ptr reference (latent bug; entry was
      already raw-released). Fix: stamp generation through the raw
      pointer.

D-P1-6: `(size_t)(-n) + 1024` — `n` is plugin-returned int32. If
n == INT_MIN, `-n` is signed-negation UB; previous code would either
wrap to INT_MIN (cast to a huge size_t, vector::resize throws
bad_alloc) or trigger UBSAN. Replace verbatim across all 9 call sites
with the safe `(size_t)(-(int64_t)n) + 1024`.

Files touched (all fixes are local; no API change):
  backend/include/xi/xi_ipc.hpp        Reader::str/bytes bounds check
  backend/include/xi/xi_shm.hpp        alloc_image dim guard
  backend/include/xi/xi_image_pool.hpp create dim/counter ordering
  backend/include/xi/xi_abi.hpp        UB pattern fix
  backend/include/xi/xi_plugin_manager.hpp UB pattern fix (2x)
  backend/include/xi/xi_plugin_handle.hpp  UB pattern fix (2x)
  backend/include/xi/xi_use.hpp        UB pattern fix
  backend/src/worker_main.cpp          UB pattern fix (2x)
  backend/src/service_main.cpp         UB pattern fix (5x)
  backend/src/script_runner_main.cpp   UB pattern fix

Validation:
  cross_proc_trigger PASS (60 ≥ 50 inspects)
  r6_p2_demo 7/7 PASS
  fl_r7_fuzz evil_worker 16 strategies — 0 fatals, 0 findings

+++ backend/include/xi/xi_abi.hpp
@@ -164,7 +164,7 @@ public:
         int32_t n = host_->instance_folder(name_.c_str(), buf, sizeof(buf));
         if (n > 0) return std::string(buf, (size_t)n);
         if (n < 0) {
-            std::vector<char> big((size_t)(-n) + 1);
+            std::vector<char> big((size_t)(-(int64_t)n) + 1);
             n = host_->instance_folder(name_.c_str(), big.data(), (int32_t)big.size());
             if (n > 0) return std::string(big.data(), (size_t)n);
         }
