+++ backend/include/xi/xi_plugin_handle.hpp
@@ -125,7 +125,7 @@ public:
         if (!instance_ || !exchange_fn_) return "{}";
         std::vector<char> buf(64 * 1024);
         int n = exchange_fn_(instance_, cmd_json.c_str(), buf.data(), (int)buf.size());
-        if (n < 0) { buf.resize((size_t)(-n) + 1024);
+        if (n < 0) { buf.resize((size_t)(-(int64_t)n) + 1024);
                      n = exchange_fn_(instance_, cmd_json.c_str(), buf.data(), (int)buf.size()); }
         return (n > 0) ? std::string(buf.data(), (size_t)n) : "{}";
     }
@@ -134,7 +134,7 @@ public:
         if (!instance_ || !get_def_fn_) return "{}";
         std::vector<char> buf(4096);
         int n = get_def_fn_(instance_, buf.data(), (int)buf.size());
-        if (n < 0) { buf.resize((size_t)(-n) + 1024);
+        if (n < 0) { buf.resize((size_t)(-(int64_t)n) + 1024);
                      n = get_def_fn_(instance_, buf.data(), (int)buf.size()); }
         return (n > 0) ? std::string(buf.data(), (size_t)n) : "{}";
     }
